v0.0.17
- Added Twitter for Players
- Strip FriendCode from name and return it. Fixes https://github.com/kjhf/SplatTag/issues/8
- Fixed team ids not merging when starting on 0 (NoTeam)
- Fix JsonConvert.DefaultSettings NRE
- Better logging of loading process
- All unit tests now pass again.

+++ SplatTagUnitTests/SplatTagUnitTest.cs
@@ -70,7 +70,7 @@ namespace SplatTagUnitTests
 
       object teamsDict = Util.GetPrivateMember(controller, "teams");
       Assert.IsNotNull(teamsDict);
-      var dictionary2 = (IDictionary<uint, Team>)teamsDict;
+      var dictionary2 = (IDictionary<long, Team>)teamsDict;
       Assert.IsNotNull(dictionary2);
       Assert.IsTrue(dictionary2.TryGetValue(TEAM_ID, out Team target2));
       Assert.IsTrue(target2.Id == TEAM_ID);
